OGM-1089 Rename existing integration tests for Neo4j remote
